Move run and affected under exec category
