+++ src/cli/commands/exec/run.ts
@@ -1,22 +1,22 @@
-import type { MetricStore } from "../storage/types.js";
+import type { MetricStore } from "../../storage/types.js";
 import {
   planSample,
   type SamplingSummary,
-} from "./sample.js";
+} from "../sample.js";
 import type { SamplingMode } from "./sampling-options.js";
-import type { QuarantineManifestEntry } from "../quarantine-manifest.js";
-import type { DependencyResolver } from "../resolvers/types.js";
+import type { QuarantineManifestEntry } from "../../quarantine-manifest.js";
+import type { DependencyResolver } from "../../resolvers/types.js";
 import {
   orchestrate,
   withQuarantineRuntime,
   type ExecuteResult,
   type RunnerAdapter,
   type TestId,
-} from "../runners/index.js";
+} from "../../runners/index.js";
 import {
   createMetaKey,
   buildListedTestIndex,
-} from "./test-key.js";
+} from "../test-key.js";
 
 export interface RunOpts {
   store: MetricStore;
