+++ tests/cli/full-strategy.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { parseSamplingMode } from "../../src/cli/commands/sampling-options.js";
+import { parseSamplingMode } from "../../src/cli/commands/exec/sampling-options.js";
 
 describe("full sampling mode", () => {
   it("parseSamplingMode accepts 'full'", () => {
